Split effects from cog
